[content] fixed bug where profile didn't show
Example:
  - go to your local ingress community
  - click a user
  - click 'About'
What do you expect:
  A card showing the ingress agent data
What do you see:
  No card

Reason of the bug:
When changing tabs from posts to about, the about page gets
rebuilt by G+. The parent element doesn't, and that's the
element we used to timestamp the creation.
Now we moved to a system where the parent is used if the user
is not an Ingress player, and the element itself when he is.